// src/app/models/orders.ts
export class Orders {
    order_id: number;
    order_name: string = "";
    order_official_name: string = "";
    order_summary: string = "";
    order_description: string = "";
    order_banner: string = "";
    order_scope: string = "";
    order_alignment: string = "";
    order_headquarters: string = "";
    order_values: string = "";
    order_goals: string = "";
    order_allies: string = "";
    order_enemies: string = ""; 
    order_rivals: string = ""; 
}